+++ src/greynoise/cli/formatter.py
@@ -169,6 +169,7 @@ def interesting_formatter(results, verbose):
     max_width, _ = shutil.get_terminal_size()
     return template.render(results=results, verbose=verbose, max_width=max_width)
 
+
 @colored_output
 def similar_formatter(results, verbose):
     """Convert IP Sim to human-readable text."""
@@ -176,10 +177,19 @@ def similar_formatter(results, verbose):
     max_width, _ = shutil.get_terminal_size()
     return template.render(results=results, verbose=verbose, max_width=max_width)
 
+
 @colored_output
 def timeline_formatter(results, verbose):
     """Convert IP Sim to human-readable text."""
     template = JINJA2_ENV.get_template("timeline.txt.j2")
+    max_width, _ = shutil.get_terminal_size()
+    return template.render(results=results, verbose=verbose, max_width=max_width)
+
+
+@colored_output
+def timelinehourly_formatter(results, verbose):
+    """Convert IP Sim to human-readable text."""
+    template = JINJA2_ENV.get_template("timelinehourly.txt.j2")
     max_width, _ = shutil.get_terminal_size()
     return template.render(results=results, verbose=verbose, max_width=max_width)
 
@@ -198,5 +208,6 @@ FORMATTERS = {
         "ip-multi": ip_multi_context_formatter,
         "similar": similar_formatter,
         "timeline": timeline_formatter,
+        "timelinehourly": timelinehourly_formatter
     },
 }